back/requirements.txt
asgiref==3.5.0
astroid==2.11.2
autopep8==1.6.0
bcrypt==4.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dill==0.3.4
dj-rest-auth==2.2.4
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysqlclient==2.1.1
oauthlib==3.2.0
paramiko==2.11.0
pep8==1.7.1
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.13.4
PyNaCl==1.5.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
sshtunnel==0.4.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
wrapt==1.14.0